requirements.txt
feedparser>=6.0
python-dotenv>=1.0
openai>=1.0
requests>=2.31
schedule>=1.2
